requirements.txt
requests>=2.31.0
aiosmtplib>=2.0.0
schedule>=1.2.0
prometheus-client>=0.17.0
aiohttp>=3.8.5
backoff>=2.2.1
jsonpath-ng>=1.5.3
urllib3>=2.0.0
Flask>=2.3.0
python-dotenv
